add some comment, refactoring some code

+++ VocabNote/app/src/main/java/com/android/vocab_note/ViewModels/CategoryManagerViewModel.java
@@ -3,6 +3,7 @@ package com.android.vocab_note.ViewModels;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.ViewModel;
 
+import com.android.vocab_note.Constants;
 import com.android.vocab_note.DataRepository;
 import com.android.vocab_note.Model.Entity.Category;
 
@@ -30,19 +31,29 @@ public class CategoryManagerViewModel extends ViewModel
         return categories.get(position);
     }
 
+
+    /** Remove a category, change all the words's category int it to Common
+     * @param position the position of the category to remove
+     */
     public void removeCategory(int position)
     {
         Category removedCategory = getCategory(position);
-        boolean canDelete = repository.deleteCategory(removedCategory);
 
-        //TODO change the words in a deleted category to COMMON
+        //Cannot remove the default category
+        if(removedCategory.getCategory().equals(Constants.CATEGORY_COMMON))
+            return;
 
-        if(canDelete)
-        {
-            repository.changeWordsToDefaultCategory(removedCategory.getId());
-        }
+        repository.deleteCategory(removedCategory);
+
+        //change the words in a deleted category to COMMON
+        repository.changeWordsToDefaultCategory(removedCategory.getId());
     }
 
+
+    /** Rename a category
+     * @param oldCategory the old category
+     * @param newCategoryStr the new category
+     */
     public void renameCategory(Category oldCategory, String newCategoryStr)
     {
         Category newCategory = new Category(oldCategory.getId(), newCategoryStr);
